week02_management_and_testing/homework/requirements.txt
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
pytest==7.4.4
pytest-cov==4.1.0
wandb==0.15.0
hydra-core==1.3.2
omegaconf==2.3.0